ISPN-6572 JCache remote management access option
* Added option that allows JCache remote managament access to be
  disabled. When this option is enabled, creating or removing a cache
  throws an exception indicating that the operation is not
  possible. This option is designed for situations where the server has
  pre-defined all the user caches.
