+++ jcache/remote/src/main/java/org/infinispan/jcache/remote/JCacheManager.java
@@ -20,8 +20,6 @@ import org.infinispan.commons.util.ReflectionUtil;
 import org.infinispan.jcache.AbstractJCache;
 import org.infinispan.jcache.AbstractJCacheManager;
 import org.infinispan.jcache.logging.Log;
-import org.infinispan.jcache.remote.ServerManager.ManagementClientException;
-import org.infinispan.jcache.remote.ServerManager.NotAvailableException;
 import org.infinispan.commons.logging.LogFactory;
 
 public class JCacheManager extends AbstractJCacheManager {
@@ -30,7 +28,7 @@ public class JCacheManager extends AbstractJCacheManager {
    private RemoteCacheManager cm;
    private RemoteCacheManager cmForceReturnValue;
 
-   private ServerManager sm;
+   private RemoteCacheAccess cacheAccess;
 
    public JCacheManager(URI uri, ClassLoader classLoader, CachingProvider provider, Properties properties) {
       super(uri, classLoader, provider, properties, false);
@@ -43,7 +41,7 @@ public class JCacheManager extends AbstractJCacheManager {
       builder.forceReturnValues(true);
       cmForceReturnValue = new RemoteCacheManager(builder.build(), true);
 
-      sm = new ServerManager(configuration.servers().get(0).host());
+      cacheAccess = RemoteCacheAccess.createCacheAccess(configuration, properties);
    }
 
    private ConfigurationBuilder getConfigurationBuilder(Properties userProperties) {
@@ -82,16 +80,7 @@ public class JCacheManager extends AbstractJCacheManager {
    protected <K, V, C extends Configuration<K, V>> AbstractJCache<K, V> create(String cacheName, C configuration) {
       ConfigurationAdapter<K, V> adapter = ConfigurationAdapter.create(configuration);
 
-      try {
-         if (!sm.containsCache(cacheName)) {
-            sm.addCache(cacheName);
-         }
-      } catch (NotAvailableException ex) {
-         //Nothing to do.
-      } catch (ManagementClientException ex) {
-         throw log.cacheCreationFailed(cacheName, ex);
-      }
-
+      cacheAccess.addCacheIfAbsent(cacheName);
       return createJCache(cacheName, adapter);
    }
 
@@ -167,13 +156,7 @@ public class JCacheManager extends AbstractJCacheManager {
 
    @Override
    protected <K, V> void delegateRemoveCache(AbstractJCache<K, V> jcache) {
-      try {
-         sm.removeCache(jcache.getName());
-      } catch (NotAvailableException ex) {
-         // Nothing to do.
-      } catch (ManagementClientException ex) {
-         throw log.serverManagementOperationFailed(ex);
-      }
+      cacheAccess.removeCache(jcache.getName());
       jcache.close();
    }
